Bugfix: check the status of profile mount/umount at the end of the session
git-svn-id: file:///home/code/svn/ovd/trunk@4091 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ OvdServer/ovd/Role/ApplicationServer/Platform/Windows/Session.py
@@ -122,16 +122,18 @@ class Session(AbstractSession):
 	
 	def uninstall_client(self):
 		if self.profile is not None:
-			self.profile.mount()
-			
-			self.profile.copySessionStop()
-			
-			for shortcut in self.installedShortcut:
-				dstFile = os.path.join(self.profile.mountPoint, self.profile.DesktopDir, shortcut)
-				if os.path.exists(dstFile):
-					os.remove(dstFile)
-			
-			self.profile.umount()
+			if not self.profile.mount():
+				Logger.warn("Unable to mount profile at uninstall_client of session "+self.id)
+			else:
+				self.profile.copySessionStop()
+				
+				for shortcut in self.installedShortcut:
+					dstFile = os.path.join(self.profile.mountPoint, self.profile.DesktopDir, shortcut)
+					if os.path.exists(dstFile):
+						os.remove(dstFile)
+				
+				if not self.profile.umount():
+					Logger.error("Unable to umount profile at uninstall_client of session "+self.id)
 		
 		self.user.destroy()
